+++ tests/DateTime/DateFallback.test.ts
@@ -6,7 +6,7 @@ import moment from 'moment';
 import { TasksFile } from '../../src/Scripting/TasksFile';
 import { Task } from '../../src/Task/Task';
 import { resetSettings, updateSettings } from '../../src/Config/Settings';
-import { DateFallback } from '../../src/Task/DateFallback';
+import { DateFallback } from '../../src/DateTime/DateFallback';
 import { TaskLocation } from '../../src/Task/TaskLocation';
 import { TaskBuilder } from '../TestingTools/TaskBuilder';
